big commit(

+++ aaaSystemsCommon/Models/RoomMessage.cs
@@ -7,7 +7,7 @@ namespace aaaSystemsCommon.Models
     public class RoomMessage : IEntity
     {
         public int Id { get; set; }
-        public long MessageId { get; set; }
+        public int MessageId { get; set; }
         public long UserId { get; set; }
         public DateTime DateTime { get; set; }
         public int RoomId { get; set; }
